🚚 Move: useUserQuery로 프로필훅 이동
