+++ src/hooks/useUserQuery.js
@@ -1,6 +1,20 @@
-import { useQuery } from '@tanstack/react-query';
+import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import { fetchUserData } from '../api/userInfoService';
 import { QUERY_KEYS } from '../constants/queryKeys';
+import { useState } from 'react';
+import {
+  checkNicknameDuplication,
+  updateProfile,
+  validateNickname,
+} from '../utils/profileUtils';
+import {
+  ALLOWED_IMAGE_TYPES,
+  BUCKET_NAME,
+  MAX_FILE_SIZE,
+  PRODUCT_DEFAULT_IMG,
+  PROFILES_DIRECTORY,
+} from '../constants/mypageConstants';
+import { supabase } from '../api/client';
 
 /**
  * 특정 사용자의 데이터를 가져오는 React Query 훅
@@ -20,4 +34,132 @@ export const useUserData = (sub) => {
     },
     enabled: !!sub,
   });
+};
+
+export const useProfileInfo = (user) => {
+  const queryClient = useQueryClient();
+  const [localError, setLocalError] = useState(null);
+
+  // 유저 데이터 조회
+  const { data: userdata } = useQuery({
+    queryKey: ['user', user?.id], //QUERY_KEYS.USER를 하면 적용이 안됨
+    queryFn: () => fetchUserData(user.id),
+    enabled: !!user?.id,
+    select: (data) => data.data,
+  });
+
+  // 프로필 업데이트
+  const updateProfileMutation = useMutation({
+    mutationFn: (newNickname) => updateProfile(newNickname, userdata),
+    onSuccess: () => {
+      queryClient.invalidateQueries(['user', user?.id]);
+    },
+  });
+
+  // 닉네임 중복 검사
+  const checkNicknameMutation = useMutation({
+    mutationFn: checkNicknameDuplication,
+  });
+
+  // 프로필 업데이트 핸들러
+  const handleProfileUpdate = async (nickname, isUpdating) => {
+    if (!isUpdating) return { isUpdating: true };
+
+    setLocalError(null);
+
+    // 유효성 검사
+    const validation = validateNickname(nickname, userdata?.name);
+    if (!validation.valid) {
+      if (validation.isUnchanged) {
+        return { unchanged: true };
+      }
+      setLocalError(new Error(validation.error));
+      return;
+    }
+
+    // 닉네임 중복 검사
+    try {
+      await checkNicknameMutation.mutateAsync(nickname);
+    } catch (error) {
+      setLocalError(error);
+      return;
+    }
+
+    // 프로필 업데이트
+    try {
+      await updateProfileMutation.mutateAsync(nickname);
+    } catch (error) {
+      setLocalError(error);
+      return;
+    }
+
+    return { success: true };
+  };
+
+  return {
+    userdata,
+    handleProfileUpdate,
+    error:
+      localError || checkNicknameMutation.error || updateProfileMutation.error,
+  };
+};
+
+export const useProfileImage = (userdata) => {
+  const queryClient = useQueryClient();
+
+  const uploadImageMutation = useMutation({
+    mutationFn: async (file) => {
+      const fileName = `${Date.now()}-${Math.floor(Math.random() * 100000)}.${file.name.split('.').pop()}`;
+      const filePath = `${PROFILES_DIRECTORY}/${fileName}`;
+
+      const { error } = await supabase.storage
+        .from(BUCKET_NAME)
+        .upload(filePath, file, { upsert: false });
+
+      if (error) throw error;
+
+      const { data: urlData } = await supabase.storage
+        .from(BUCKET_NAME)
+        .getPublicUrl(filePath);
+
+      return urlData.publicUrl;
+    },
+  });
+
+  const updateProfileImageMutation = useMutation({
+    mutationFn: async (imageUrl) => {
+      const { error } = await supabase
+        .from('users')
+        .update({ profile_img: imageUrl })
+        .eq('user_id', userdata.user_id);
+
+      if (error) throw error;
+    },
+    onSuccess: () => {
+      queryClient.invalidateQueries(['user', userdata?.user_id]);
+    },
+  });
+
+  const handleImageChange = async (e) => {
+    const file = e.target.files[0];
+    if (!file) return;
+
+    if (file.size > MAX_FILE_SIZE) {
+      alert('2MB 이하의 파일을 선택해주세요.');
+    }
+    if (!ALLOWED_IMAGE_TYPES.includes(file.type)) {
+      alert('JPG 또는 PNG 형식의 이미지만 업로드 가능합니다.');
+    }
+
+    uploadImageMutation.mutate(file, {
+      onSuccess: (imageUrl) => {
+        updateProfileImageMutation.mutate(imageUrl);
+      },
+    });
+  };
+
+  return {
+    imageUrl: userdata?.profile_img || PRODUCT_DEFAULT_IMG,
+    handleImageChange,
+  };
 };
